move testdata factories from types to utils
