+++ wls-gui-wahllokalsystem/tests/composables/broadcast/broadcastService.spec.ts
@@ -1,4 +1,4 @@
-import { useBroadcastTestDataFactory } from "@tests/types/broadcast/BroadcastTestDataFactory.ts";
+import { useBroadcastTestDataFactory } from "@tests/utils/broadcast/BroadcastTestDataFactory.ts";
 import { beforeEach, describe, expect, it, vi } from "vitest";
 
 import { useBroadcastService } from "@/composables/broadcast/broadcastService.ts";
